requirements.txt
Flask==2.1.0         
Flask-Cors==3.0.10         
Flask-SQLAlchemy==2.5.1
APScheduler==3.9.1
spotipy==2.19.0              
SQLAlchemy==1.4.32  
python-vlc==3.0.18122